Safari-611.1.10/LayoutTests/fast/canvas/canvas-ellipse-360-winding-expected.txt

This tests canvas full arc fill with nonzero winding rule. Eight green concentric thick circumferences should be displayed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS data[1] is 0
PASS data[1] is 255
PASS data[1] is 255
PASS data[1] is 0
PASS data[1] is 0
PASS data[1] is 255
PASS successfullyParsed is true

TEST COMPLETE
